+++ src/parser/addressRepresentationConverters.ts
@@ -3,19 +3,20 @@
  * Copyright (c) 2021 Handsoncode. All rights reserved.
  */
 
+import {simpleCellRange, SimpleCellRange} from '../AbsoluteCellRange'
 import {simpleCellAddress, SimpleCellAddress} from '../Cell'
 import {Maybe} from '../Maybe'
 import {CellAddress} from './CellAddress'
 import {ColumnAddress} from './ColumnAddress'
-import {sheetNameRegexp, simpleSheetName} from './LexerConfig'
+import {ABSOLUTE_OPERATOR, RANGE_OPERATOR, sheetNameRegexp, simpleSheetName} from './LexerConfig'
 import {RowAddress} from './RowAddress'
 
 export type SheetMappingFn = (sheetName: string) => Maybe<number>
 export type SheetIndexMappingFn = (sheetIndex: number) => Maybe<string>
 
-const addressRegex = new RegExp(`^(${sheetNameRegexp})?(\\$?)([A-Za-z]+)(\\$?)([0-9]+)$`)
-const columnRegex = new RegExp(`^(${sheetNameRegexp})?(\\$?)([A-Za-z]+)$`)
-const rowRegex = new RegExp(`^(${sheetNameRegexp})?(\\$?)([0-9]+)$`)
+const addressRegex = new RegExp(`^(${sheetNameRegexp})?(\\${ABSOLUTE_OPERATOR}?)([A-Za-z]+)(\\${ABSOLUTE_OPERATOR}?)([0-9]+)$`)
+const columnRegex = new RegExp(`^(${sheetNameRegexp})?(\\${ABSOLUTE_OPERATOR}?)([A-Za-z]+)$`)
+const rowRegex = new RegExp(`^(${sheetNameRegexp})?(\\${ABSOLUTE_OPERATOR}?)([0-9]+)$`)
 const simpleSheetNameRegex = new RegExp(`^${simpleSheetName}$`)
 
 /**
@@ -41,11 +42,11 @@ export const cellAddressFromString = (sheetMapping: SheetMappingFn, stringAddres
   }
 
   const row = Number(result[8]) - 1
-  if (result[5] === '$' && result[7] === '$') {
+  if (result[5] === ABSOLUTE_OPERATOR && result[7] === ABSOLUTE_OPERATOR) {
     return CellAddress.absolute(col, row, sheet)
-  } else if (result[5] === '$') {
+  } else if (result[5] === ABSOLUTE_OPERATOR) {
     return CellAddress.absoluteCol( col, row - baseAddress.row, sheet)
-  } else if (result[7] === '$') {
+  } else if (result[7] === ABSOLUTE_OPERATOR) {
     return CellAddress.absoluteRow(col - baseAddress.col, row, sheet)
   } else {
     return CellAddress.relative(row - baseAddress.row, col - baseAddress.col, sheet)
@@ -66,7 +67,7 @@ export const columnAddressFromString = (sheetMapping: SheetMappingFn, stringAddr
 
   const col = columnLabelToIndex(result[6])
 
-  if (result[5] === '$') {
+  if (result[5] === ABSOLUTE_OPERATOR) {
     return ColumnAddress.absolute(col, sheet)
   } else {
     return ColumnAddress.relative(col - baseAddress.col, sheet)
@@ -87,7 +88,7 @@ export const rowAddressFromString = (sheetMapping: SheetMappingFn, stringAddress
 
   const row = Number(result[6]) - 1
 
-  if (result[5] === '$') {
+  if (result[5] === ABSOLUTE_OPERATOR) {
     return RowAddress.absolute(row, sheet)
   } else {
     return RowAddress.relative(row - baseAddress.row, sheet)
@@ -122,6 +123,26 @@ export const simpleCellAddressFromString = (sheetMapping: SheetMappingFn, string
   return simpleCellAddress(sheet, col, row)
 }
 
+export const simpleCellRangeFromString = (sheetMapping: SheetMappingFn, stringAddress: string, sheetContext: number): Maybe<SimpleCellRange> => {
+  const split = stringAddress.split(RANGE_OPERATOR)
+  if(split.length !== 2) {
+    return undefined
+  }
+  const [startString, endString] = split
+  const start = simpleCellAddressFromString(sheetMapping, startString, sheetContext)
+  if(start === undefined) {
+    return undefined
+  }
+  const end = simpleCellAddressFromString(sheetMapping, endString, start.sheet)
+  if(end === undefined) {
+    return undefined
+  }
+  if(start.sheet !== end.sheet) {
+    return undefined
+  }
+  return simpleCellRange(start, end)
+}
+
 /**
  * Returns string representation of absolute address
  * If sheet index is not present in sheet mapping, returns undefined
@@ -142,6 +163,16 @@ export const simpleCellAddressToString = (sheetIndexMapping: SheetIndexMappingFn
     return `${sheetName}!${column}${address.row + 1}`
   } else {
     return `${column}${address.row + 1}`
+  }
+}
+
+export const simpleCellRangeToString = (sheetIndexMapping: SheetIndexMappingFn, address: SimpleCellRange, sheetIndex: number): Maybe<string> => {
+  const startString = simpleCellAddressToString(sheetIndexMapping, address.start, sheetIndex)
+  const endString = simpleCellAddressToString(sheetIndexMapping, address.end, address.start.sheet)
+  if(startString === undefined || endString === undefined) {
+    return undefined
+  } else {
+    return `${startString}${RANGE_OPERATOR}${endString}`
   }
 }