+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added parsing of arrays in formulas (together with respective config options for separators). (#671)
 - Added utility function for filling ranges with source from other range. (#678)
 - Added pretty print for detailedCellError. (#712)
+- Added `simpleCellRangeFromString` and `simpleCellRangeToString` helpers. (#720)
 
 ### Fixed
 - Fixed an issue with arrays and cruds. (#651)